added files!
this is a simple project to support my web designing process where I generally spend quite a bit of time for creating a palette. this site is responsive.
although this is made for fun! this lacks accessbiblity and also there are some bugs that are to be fixed in javascript file!